+++ src/main/java/io/github/artsobol/bookcrossing/feature/author/service/AuthorFinder.java
@@ -1,0 +1,8 @@
+package io.github.artsobol.bookcrossing.feature.author.service;
+
+import io.github.artsobol.bookcrossing.feature.author.entity.Author;
+
+public interface AuthorFinder {
+
+    Author findById(Long id);
+}