+++ src/main/java/io/github/artsobol/bookcrossing/feature/book/dto/UpdateBookRequest.java
@@ -1,0 +1,9 @@
+package io.github.artsobol.bookcrossing.feature.book.dto;
+
+public record UpdateBookRequest(
+        String title,
+        String description,
+        Long authorId,
+        Long genreId
+) {
+}