+++ src/main/java/io/github/artsobol/bookcrossing/feature/book/entity/BookStatus.java
@@ -1,0 +1,7 @@
+package io.github.artsobol.bookcrossing.feature.book.entity;
+
+public enum BookStatus {
+    AVAILABLE,
+    RESERVED,
+    EXCHANGED
+}