+++ src/main/java/io/github/artsobol/bookcrossing/feature/genre/service/GenreFinder.java
@@ -1,0 +1,8 @@
+package io.github.artsobol.bookcrossing.feature.genre.service;
+
+import io.github.artsobol.bookcrossing.feature.genre.entity.Genre;
+
+public interface GenreFinder {
+
+    Genre findById(Long id);
+}